requirements.txt
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.12
DateTime==4.4
geographiclib==1.52
geopy==2.2.0
gmplot==1.4.1
greenlet==1.1.2
idna==3.3
kaleido==0.2.1
numpy==1.22.3
oauthlib==3.2.0
pandas==1.4.2
plotly==5.7.0
psycopg2-binary==2.9.3
pycodestyle==2.8.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
SQLAlchemy==1.4.36
tenacity==8.0.1
toml==0.10.2
tweepy==4.9.0
urllib3==1.26.9
zope.interface==5.4.0
